requirements.txt
cachetools==3.1.1
Django==2.2.4
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
numpy==1.17.0
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==2.2.2
pyasn1==0.4.6
pyasn1-modules==0.2.6
pytesseract==0.2.9
python-dateutil==2.8.0
pytz==2019.2
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
xlrd==1.2.0
